gRPC and VGen tasks for pipeline
Adds tasks for running gRPC and Veneer codegen to the pipeline, as
well as a minimal, serial flow to drive the tasks. Also adds setup
information to the README.

Note that this uses the revised version of CodeGeneratorTool in
c/1562, which changes the command line arguments to the tool. That CL
has not yet been submitted, so you must patch that change if you want
to run the flow.

Change-Id: Ic75905714374f6ff49243e229624145608233e35