add github push task, using util function in pipeline/utils.  (#50)
add github push task, using util function in pipeline/utils.

Once we had extended the Python and Java GRPC client pipelines in artman
to optionally publish generated code, we wanted to extend the Go
pipeline to do the same. For Go the standard practice for publishing is
simply to upload code to a GitHub repository. Since the task of
uploading generated code to a GitHub repository might be useful for
other languages as well, we want to add a task that will upload a folder
to a GitHub repository (independent of the particular language). Code
similar to this will also be used for publishing API protos from google3
to GitHub.

This commit uses the GitHub API's JSON over HTTP methods to create and
push the commit, because libraries that wrap the GitHub API did not
provide the required methods. Using shell commands as with maven and
pypi would be more complicated in this case since we would need to
clone/add/commit/push instead of running a single publish command. Using
the GitHub API gives better error handling and allows us to reuse the
code for gapic projects where we won't have access to the git binary.

The effect of running the task is not exactly the same as that of a
forced git push: this won't change or delete files that are missing
in the local repo but are present in the remote repo. This is 
another reason to use the GitHub API rather than shell commands.

The implementation follows these steps (suggested on the API page),
manipulating git internals to publish the output folder:
1. get the current branch and its commit object
2. retrieve the tree that the commit points to
3. post a new tree object with changed or new files, with the old tree
as a parent
4. create a new commit object with the current commit SHA as the parent
and the new tree SHA, getting a commit SHA back
5. update the reference of the  branch to point to the new commit SHA