Update artman to use language-independent pipelines (#61)
This PR removes the language-specific pipelines (e.g. PythonGapicClientPipeline, JavaGapicClientPipeline, etc.) and replaces them with language-independent pipelines (e.g. GapicClientPipeline) and a command line flag to specify the language.

This results in a simplified command line interface, with the language having to be specified in exactly one place (the --language argument) rather than as part of the pipeline name and appended to the config files.