Change naming configuration (#123)
Updates #116

Currently api_name is parsed to derive many smaller pieces of information:

- a one-word short name
- a major version
- a path to the source protos in the googleapis repo for package metadata
- the name of the package

Instead of parsing this information out, it makes more sense to configure these values separately. This avoids corner cases where the short name may need to consist of multiple words (e.g., error-reporting, iam-admin), or where the name of the package may not match the path to the protos in the googleapis repo.

This PR separates out the short_name and version configuration. A subsequent PR will replace the source to the protos currently parsed from the api_name with the existing src_proto_path configuration value.