Don't dump Python classes in package metadata YAML (#135)
A user running remote artman encountered a failure where artman dumped
a `!!python/unicode 'v1beta1'` object in the package metadata. This
caused an error when the `toolkit` Java code tried to load the YAML.

Use `safe_dump` instead of `dump` to ensure that Python objects are
not being dumped in the package metadata.