Fix non-master baseline branch bug. (#180)
This commit fixes a logic error that cropped up if the baseline
branch is something other than master. Essentially, if the baseline
branch was out of date (common), then the resulting PR would have
all of the commits from master which are missing, because it
always created the new branch off of master, even if master is
not the target.

This ensures that the baseline branch is the actual baseline for
creating the new branch.