Introduce new artman config spec and artman CLI (#254)
Introduce a new artman config spec and artman CLI, so that the artman spec is formalized from the artman user perspective. This PR doesn't change the implementation of artman in-memory data model which remains as a dict. This PR parses an artman yaml in the new spec and turn it into a legacy yaml, so that the spec change becomes transparent to artman internal execution engine. A follow-up PR will make another round of refactoring that switches to use the new artman config model (a protobuf class) instead of the legacy config dict.

artman2 will by default executes inside an artman Docker container, unless --local is passed. Once all current artman config yaml switch to the new config spec, artman2 will be renamed to artman and the legacy artman will be removed. And artman version will become 0.5.*.

To prove the correctness of the conversion, test_converter.py is a unittest to compare the two formats once loaded into artman. In addition to that, artman golden test checks the output using the new and old spec/CLI , and make sure the same set of output gets generated.

For details of the spec and CLI syntax, please refer to the tables in https://github.com/googleapis/artman/pull/254