wrap all parameters in 'apostrophes' (#656)
When we build a Docker command line, we should be careful when our parameters contain spaces.
E.g. if we want to pass `--generator-args="arg1 arg2"`, it's important to wrap _this whole_ string in apostrophes or quotes, otherwise the inner Artman in docker will receive it as `--generator-args=arg1 arg2` and will fail. This PR makes it `'--generator-args=arg1 arg2'` which works fine.