DiscogapicConfigGen: Use a temporary output directory (#669)
DiscogapicConfigGen generates gapic yaml from discovery docs, and therefore the flag gapic_yaml is used as output instead of input (which is the usual case for codegen).

When trying to refresh gapic config for compute artman creates this default artman-genfiles in my local environment while does not drop any output there. Therefore it seems clearer to just use a temporary directory to hold the generated gapic config before moving to the right place.

Also, if user specifies the output_dir flag when invoking artman generate discogapic_config, throw a command line warning saying that the flag is ignored, and output is saved in the directory given by gapic_yaml.